requirements.txt
flask==3.1.0
numpy==2.2.4
pandas==2.2.3
gunicorn==23.0.0
python-dotenv==1.0.1